+++ src/array/array_sync_writable.rs
@@ -216,8 +216,7 @@ impl<TStorage: ?Sized + WritableStorageTraits + 'static> Array<TStorage> {
             ));
         }
 
-        let all_fill_value = self.fill_value().equals_all(&chunk_bytes);
-        if all_fill_value {
+        if !options.store_empty_chunks() && self.fill_value().equals_all(&chunk_bytes) {
             self.erase_chunk(chunk_indices)?;
             Ok(())
         } else {