+++ src/array/codec/options.rs
@@ -3,9 +3,12 @@
 use crate::config::global_config;
 
 /// Codec options for encoding/decoding.
+///
+/// Default values for these options are set by the global [`Config`](crate::config::Config).
 #[derive(Debug, Clone)]
 pub struct CodecOptions {
     validate_checksums: bool,
+    store_empty_chunks: bool,
     concurrent_target: usize,
 }
 
@@ -13,6 +16,7 @@ impl Default for CodecOptions {
     fn default() -> Self {
         Self {
             validate_checksums: global_config().validate_checksums(),
+            store_empty_chunks: global_config().store_empty_chunks(),
             concurrent_target: global_config().codec_concurrent_target(),
         }
     }
@@ -30,6 +34,7 @@ impl CodecOptions {
     pub fn into_builder(&self) -> CodecOptionsBuilder {
         CodecOptionsBuilder {
             validate_checksums: self.validate_checksums,
+            store_empty_chunks: self.store_empty_chunks,
             concurrent_target: self.concurrent_target,
         }
     }
@@ -45,6 +50,17 @@ impl CodecOptions {
         self.validate_checksums = validate_checksums;
     }
 
+    /// Return the store empty chunks setting.
+    #[must_use]
+    pub fn store_empty_chunks(&self) -> bool {
+        self.store_empty_chunks
+    }
+
+    /// Set whether or not to store empty chunks.
+    pub fn set_store_empty_chunks(&mut self, store_empty_chunks: bool) {
+        self.store_empty_chunks = store_empty_chunks;
+    }
+
     /// Return the concurrent target.
     #[must_use]
     pub fn concurrent_target(&self) -> usize {
@@ -58,9 +74,12 @@ impl CodecOptions {
 }
 
 /// Builder for [`CodecOptions`].
+///
+/// Default values for these options are set by the global [`Config`](crate::config::Config).
 #[derive(Debug, Clone)]
 pub struct CodecOptionsBuilder {
     validate_checksums: bool,
+    store_empty_chunks: bool,
     concurrent_target: usize,
 }
 
@@ -76,6 +95,7 @@ impl CodecOptionsBuilder {
     pub fn new() -> Self {
         Self {
             validate_checksums: global_config().validate_checksums(),
+            store_empty_chunks: global_config().store_empty_chunks(),
             concurrent_target: global_config().codec_concurrent_target(),
         }
     }
@@ -85,6 +105,7 @@ impl CodecOptionsBuilder {
     pub fn build(&self) -> CodecOptions {
         CodecOptions {
             validate_checksums: self.validate_checksums,
+            store_empty_chunks: self.store_empty_chunks,
             concurrent_target: self.concurrent_target,
         }
     }
@@ -93,6 +114,13 @@ impl CodecOptionsBuilder {
     #[must_use]
     pub fn validate_checksums(mut self, validate_checksums: bool) -> Self {
         self.validate_checksums = validate_checksums;
+        self
+    }
+
+    /// Set whether or not to store empty chunks.
+    #[must_use]
+    pub fn store_empty_chunks(mut self, store_empty_chunks: bool) -> Self {
+        self.store_empty_chunks = store_empty_chunks;
         self
     }